Move Ipv4Address to data, import SavedObjectNotFound from common.
