+++ src/plugins/data/common/search/aggs/ipv4_address.test.ts
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-// @ts-ignore
+
 import expect from '@kbn/expect';
 import { Ipv4Address } from './ipv4_address';
