+++ src/plugins/data/public/index.ts
@@ -306,6 +306,7 @@ import {
   dateHistogramInterval,
   InvalidEsCalendarIntervalError,
   InvalidEsIntervalFormatError,
+  Ipv4Address,
   isValidEsInterval,
   isValidInterval,
   parseEsInterval,
@@ -373,6 +374,7 @@ export const search = {
     intervalOptions,
     InvalidEsCalendarIntervalError,
     InvalidEsIntervalFormatError,
+    Ipv4Address,
     isDateHistogramBucketAggConfig, // TODO: remove in build_pipeline refactor
     isNumberType,
     isStringType,
@@ -417,7 +419,6 @@ export {
   connectToQueryState,
   syncQueryStateWithUrl,
   QueryState,
-  getQueryLog,
   getDefaultQuery,
   FilterManager,
   SavedQuery,
