+++ src/plugins/data/server/index.ts
@@ -155,6 +155,7 @@ import {
   dateHistogramInterval,
   InvalidEsCalendarIntervalError,
   InvalidEsIntervalFormatError,
+  Ipv4Address,
   isValidEsInterval,
   isValidInterval,
   parseEsInterval,
@@ -184,6 +185,7 @@ export const search = {
     dateHistogramInterval,
     InvalidEsCalendarIntervalError,
     InvalidEsIntervalFormatError,
+    Ipv4Address,
     isValidEsInterval,
     isValidInterval,
     parseEsInterval,
