+++ src/plugins/kibana_utils/public/index.ts
@@ -35,7 +35,6 @@ export {
 export * from './core';
 export * from '../common/errors';
 export * from './field_wildcard';
-export * from './parse';
 export * from './render_complete';
 export * from './resize_checker';
 export * from '../common/state_containers';
